refactor: rewrite defineVaporAsyncComponent import to defineAsyncComponent in SSR + vapor mode
This limits the defineVaporAsyncComponent only used in the vue file
revert https://github.com/vuejs/core/pull/13383/commits/897404317ef749c12d59f9731d0e8d94b013c6ac
