+++ packages/vue/src/index-with-vapor.ts
@@ -1,3 +1,3 @@
 // for type generation only
-export * from './indexBase'
+export * from './index'
 export * from '@vue/runtime-vapor'
